+++ packages/algolia-experiences/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.11](https://github.com/algolia/instantsearch/compare/algolia-experiences@1.8.10...algolia-experiences@1.8.11) (2026-06-19)
+
+**Note:** Version bump only for package algolia-experiences
+
+
+
+
+
+## [1.8.10](https://github.com/algolia/instantsearch/compare/algolia-experiences@1.8.9...algolia-experiences@1.8.10) (2026-06-09)
+
+**Note:** Version bump only for package algolia-experiences
+
+
+
+
+
+## [1.8.9](https://github.com/algolia/instantsearch/compare/algolia-experiences@1.8.8...algolia-experiences@1.8.9) (2026-06-03)
+
+**Note:** Version bump only for package algolia-experiences
+
+
+
+
+
+## [1.8.8](https://github.com/algolia/instantsearch/compare/algolia-experiences@1.8.7...algolia-experiences@1.8.8) (2026-06-02)
+
+**Note:** Version bump only for package algolia-experiences
+
+
+
+
+
+## [1.8.7](https://github.com/algolia/instantsearch/compare/algolia-experiences@1.8.6...algolia-experiences@1.8.7) (2026-05-26)
+
+**Note:** Version bump only for package algolia-experiences
+
+
+
+
+
 ## [1.8.6](https://github.com/algolia/instantsearch/compare/algolia-experiences@1.8.5...algolia-experiences@1.8.6) (2026-05-19)
 
 **Note:** Version bump only for package algolia-experiences